feat(cli): nest existing commands under the 'deploy' subcommand
As we expand this utility we've discussed adding other commands
unrelated to the deployment operations so move the commands under
'deploy' now to future proof us.
